Cities -> Deletion -> Added validation and test for deleting non existant city

+++ src/main/java/nl/fontys/s3/rentride_be/business/impl/city/DeleteCityUseCaseImpl.java
@@ -1,6 +1,7 @@
 package nl.fontys.s3.rentride_be.business.impl.city;
 
 import lombok.AllArgsConstructor;
+import nl.fontys.s3.rentride_be.business.exception.NotFoundException;
 import nl.fontys.s3.rentride_be.business.useCases.city.DeleteCityUseCase;
 import nl.fontys.s3.rentride_be.persistance.CityRepository;
 import org.springframework.stereotype.Service;
@@ -12,6 +13,9 @@ public class DeleteCityUseCaseImpl implements DeleteCityUseCase {
 
     @Override
     public void deleteCity(Long cityId) {
+        if (!cityRepository.existsById(cityId)) {
+            throw new NotFoundException("Delete->City");
+        }
         this.cityRepository.deleteById(cityId);
     }
 }
